+++ Demo/src/main/java/com/braintreepayments/demo/VenmoFragment.java
@@ -18,7 +18,7 @@ import com.braintreepayments.api.BraintreeClient;
 import com.braintreepayments.api.Configuration;
 import com.braintreepayments.api.ConfigurationCallback;
 import com.braintreepayments.api.VenmoAccountNonce;
-import com.braintreepayments.api.VenmoAuthorizeAccountCallback;
+import com.braintreepayments.api.VenmoTokenizeAccountCallback;
 import com.braintreepayments.api.VenmoClient;
 import com.braintreepayments.api.VenmoOnActivityResultCallback;
 import com.braintreepayments.api.VenmoRequest;
@@ -80,7 +80,7 @@ public class VenmoFragment extends BaseFragment {
                                VenmoRequest venmoRequest = new VenmoRequest()
                                        .profileId(null)
                                        .shouldVault(shouldVault);
-                               venmoClient.tokenizeVenmoAccount(activity, venmoRequest, new VenmoAuthorizeAccountCallback() {
+                               venmoClient.tokenizeVenmoAccount(activity, venmoRequest, new VenmoTokenizeAccountCallback() {
                                    @Override
                                    public void onResult(Exception error) {
                                        if (error != null) {